Fast implementation of security for tasks editing
Task managers can only update non-published tasks.

+++ src/main/java/me/nizheg/telegram/bot/chgk/service/impl/AnswerServiceImpl.java
@@ -4,21 +4,23 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+import lombok.RequiredArgsConstructor;
 import me.nizheg.telegram.bot.chgk.dto.Answer;
+import me.nizheg.telegram.bot.chgk.dto.LightTask;
+import me.nizheg.telegram.bot.chgk.exception.OperationForbiddenException;
 import me.nizheg.telegram.bot.chgk.repository.AnswerDao;
+import me.nizheg.telegram.bot.chgk.repository.TaskDao;
 import me.nizheg.telegram.bot.chgk.service.AnswerService;
 
 /**
  * @author Nikolay Zhegalin
  */
 @Service
+@RequiredArgsConstructor
 public class AnswerServiceImpl implements AnswerService {
 
     private final AnswerDao answerDao;
-
-    public AnswerServiceImpl(AnswerDao answerDao) {
-        this.answerDao = answerDao;
-    }
+    private final TaskDao taskDao;
 
     @Override
     public Answer create(Answer answer) {
@@ -32,12 +34,22 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public Answer update(Answer answer) {
+        checkPermissions(answer.getId());
         return answerDao.update(answer);
     }
 
     @Override
     public void delete(Long id) {
+        checkPermissions(id);
         answerDao.delete(id);
+    }
+
+    private void checkPermissions(Long answerId) {
+        Answer savedAnswer = answerDao.read(answerId);
+        LightTask savedTask = taskDao.getById(savedAnswer.getTaskId());
+        if (savedTask != null && savedTask.getStatus() == LightTask.Status.PUBLISHED) {
+            throw new OperationForbiddenException("It is forbidden to change answers of task in PUBLISHED status");
+        }
     }
 
     @Override
