+++ src/main/java/me/nizheg/telegram/bot/chgk/service/impl/PictureServiceImpl.java
@@ -5,17 +5,21 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+import lombok.RequiredArgsConstructor;
 import me.nizheg.telegram.bot.chgk.dto.AttachedPicture;
+import me.nizheg.telegram.bot.chgk.dto.LightTask;
 import me.nizheg.telegram.bot.chgk.dto.Picture;
+import me.nizheg.telegram.bot.chgk.exception.OperationForbiddenException;
 import me.nizheg.telegram.bot.chgk.repository.PictureDao;
+import me.nizheg.telegram.bot.chgk.repository.TaskDao;
 import me.nizheg.telegram.bot.chgk.service.PictureService;
 
 @Service
+@RequiredArgsConstructor
 public class PictureServiceImpl implements PictureService {
 
     private final PictureDao pictureDao;
-
-    public PictureServiceImpl(PictureDao pictureDao) {this.pictureDao = pictureDao;}
+    private final TaskDao taskDao;
 
     @Override
     public Picture create(Picture picture) {
@@ -29,11 +33,13 @@ public class PictureServiceImpl implements PictureService {
 
     @Override
     public Picture update(Picture picture) {
+        //todo: check permissions
         return pictureDao.update(picture);
     }
 
     @Override
     public void savePictureToTaskTextAtPosition(Long pictureId, Long taskId, int position) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.savePictureToTaskTextAtPosition(pictureId, taskId, position);
 
     }
@@ -46,6 +52,7 @@ public class PictureServiceImpl implements PictureService {
     @Override
     @Transactional
     public AttachedPicture updatePictureOfTaskText(AttachedPicture picture, Long taskId) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.update(picture);
         pictureDao.updatePictureTaskTextPosition(picture.getId(), taskId, picture.getPosition());
         return picture;
@@ -54,6 +61,7 @@ public class PictureServiceImpl implements PictureService {
     @Override
     @Transactional
     public void deletePictureFromTaskText(Long taskId, Long pictureId) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.deletePictureFromTaskText(taskId, pictureId);
         if (!pictureDao.hasPictureLinks(pictureId)) {
             pictureDao.delete(pictureId);
@@ -62,6 +70,7 @@ public class PictureServiceImpl implements PictureService {
 
     @Override
     public void savePictureToTaskCommentAtPosition(Long pictureId, Long taskId, int position) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.savePictureToTaskCommentAtPosition(pictureId, taskId, position);
 
     }
@@ -74,6 +83,7 @@ public class PictureServiceImpl implements PictureService {
     @Override
     @Transactional
     public AttachedPicture updatePictureOfTaskComment(AttachedPicture picture, Long taskId) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.update(picture);
         pictureDao.updatePictureTaskCommentPosition(picture.getId(), taskId, picture.getPosition());
         return picture;
@@ -82,9 +92,17 @@ public class PictureServiceImpl implements PictureService {
     @Override
     @Transactional
     public void deletePictureFromTaskComment(Long taskId, Long pictureId) {
+        checkPermissionsForTaskId(taskId);
         pictureDao.deletePictureFromTaskComment(taskId, pictureId);
         if (!pictureDao.hasPictureLinks(pictureId)) {
             pictureDao.delete(pictureId);
+        }
+    }
+
+    private void checkPermissionsForTaskId(Long taskId) {
+        LightTask savedTask = taskDao.getById(taskId);
+        if (savedTask != null && savedTask.getStatus() == LightTask.Status.PUBLISHED) {
+            throw new OperationForbiddenException("It is forbidden to change answers of task in PUBLISHED status");
         }
     }
